Address review suggestions: splitting jax_extras into several modules
