+++ frontend/catalyst/jax_extras/tracing.py
@@ -1,4 +1,4 @@
-# Copyright 2023 Xanadu Quantum Technologies Inc.
+# Copyright 2024 Xanadu Quantum Technologies Inc.
 
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,8 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""This module isolates utility functions that depend on JAX low-level internals
-"""
+""" Jax extras module containing functions related to the Python program tracing  """
+
 from __future__ import annotations
 
 from contextlib import ExitStack, contextmanager
@@ -35,9 +35,7 @@ from typing import (
 import jax
 from jax import ShapeDtypeStruct
 from jax._src.core import DBIdx, _update_thread_local_jit_state
-from jax._src.dispatch import jaxpr_replicas
-from jax._src.effects import ordered_effects as jax_ordered_effects
-from jax._src.interpreters.mlir import _module_name_regex, register_lowering
+from jax._src.interpreters.mlir import register_lowering
 from jax._src.interpreters.partial_eval import (
     AbstractedAxesSpec,
     _input_type_to_tracers,
@@ -45,24 +43,17 @@ from jax._src.interpreters.partial_eval import (
     trace_to_jaxpr_dynamic2,
 )
 from jax._src.lax.control_flow import _initial_style_jaxpr
-from jax._src.lax.lax import _abstractify, xla
+from jax._src.lax.lax import _abstractify
 from jax._src.lax.slicing import (
     _argnum_weak_type,
     _gather_dtype_rule,
     _gather_lower,
-    _gather_shape_computation,
-    _is_sorted,
-    _no_duplicate_dims,
-    _rank,
-    _sorted_dims_in_range,
     standard_primitive,
 )
 from jax._src.linear_util import annotate
 from jax._src.pjit import _extract_implicit_args, _flat_axes_specs
-from jax._src.sharding_impls import ReplicaAxisContext
 from jax._src.source_info_util import current as jax_current
-from jax._src.source_info_util import new_name_stack
-from jax._src.util import safe_map, unzip2, wrap_name, wraps
+from jax._src.util import safe_map, unzip2, wraps
 from jax.api_util import flatten_fun
 from jax.core import (
     AbstractValue,
@@ -79,21 +70,12 @@ from jax.core import (
     Primitive,
     ShapedArray,
     Trace,
-    Tracer,
     Var,
-    concrete_aval,
     eval_jaxpr,
     find_top_trace,
     gensym,
     new_jaxpr_eqn,
     thread_local_state,
-)
-from jax.interpreters.mlir import (
-    AxisContext,
-    ModuleContext,
-    ir,
-    lower_jaxpr_to_fun,
-    lowerable_effects,
 )
 from jax.interpreters.partial_eval import (
     DynamicJaxprTrace,
@@ -112,10 +94,10 @@ from jax.tree_util import (
 )
 from jaxlib.xla_extension import PyTreeRegistry
 
+from catalyst.jax_extras.patches import _gather_shape_rule_dynamic, get_aval2
 from catalyst.utils.patching import Patcher
 
 # pylint: disable=protected-access
-# pylint: disable=too-many-lines
 
 __all__ = (
     "ClosedJaxpr",
@@ -150,7 +132,6 @@ __all__ = (
     "input_type_to_tracers",
     "_extract_implicit_args",
     "output_type_to_tracers",
-    "jaxpr_to_mlir",
     "jaxpr_remove_implicit",
     "make_jaxpr_effects",
     "make_jaxpr2",
@@ -403,6 +384,7 @@ def deduce_avals(f: Callable, args, kwargs):
     """Wraps the callable ``f`` into a WrappedFun container accepting collapsed flatten arguments
     and returning expanded flatten results. Calculate input abstract values and output_tree promise.
     The promise must be called after the resulting wrapped function is evaluated."""
+    # TODO: deprecate in favor of `deduce_signatures`
     flat_args, in_tree = tree_flatten((args, kwargs))
     abstracted_axes = None
     axes_specs = _flat_axes_specs(abstracted_axes, *args, **kwargs)
@@ -412,126 +394,6 @@ def deduce_avals(f: Callable, args, kwargs):
     wff, out_tree_promise = flatten_fun(wf, in_tree)
     wffa = annotate(wff, in_type)
     return wffa, in_avals, keep_inputs, out_tree_promise
-
-
-def get_aval2(x):
-    """An extended version of `jax.core.get_aval` which also accepts AbstractValues."""
-    # TODO: remove this patch when https://github.com/google/jax/pull/18579 is merged
-    if isinstance(x, AbstractValue):
-        return x
-    elif isinstance(x, Tracer):
-        return x.aval
-    else:
-        return concrete_aval(x)
-
-
-def _no_clean_up_dead_vars(_eqn, _env, _last_used):
-    """A stub to workaround the Jax ``KeyError 'a'`` bug during the lowering of Jaxpr programs to
-    MLIR with the dynamic API enabled."""
-    return None
-
-
-def jaxpr_to_mlir(func_name, jaxpr):
-    """Lower a Jaxpr into an MLIR module.
-
-    Args:
-        func_name(str): function name
-        jaxpr(Jaxpr): Jaxpr code to lower
-
-    Returns:
-        module: the MLIR module corresponding to ``func``
-        context: the MLIR context corresponding
-    """
-
-    with Patcher(
-        (jax._src.interpreters.partial_eval, "get_aval", get_aval2),
-        (jax._src.core, "clean_up_dead_vars", _no_clean_up_dead_vars),
-    ):
-        nrep = jaxpr_replicas(jaxpr)
-        effects = jax_ordered_effects.filter_in(jaxpr.effects)
-        axis_context = ReplicaAxisContext(xla.AxisEnv(nrep, (), ()))
-        name_stack = new_name_stack(wrap_name("ok", "jit"))
-        module, context = custom_lower_jaxpr_to_module(
-            func_name="jit_" + func_name,
-            module_name=func_name,
-            jaxpr=jaxpr,
-            effects=effects,
-            platform="cpu",
-            axis_context=axis_context,
-            name_stack=name_stack,
-        )
-
-    return module, context
-
-
-# pylint: disable=too-many-arguments
-def custom_lower_jaxpr_to_module(
-    func_name: str,
-    module_name: str,
-    jaxpr: ClosedJaxpr,
-    effects,
-    platform: str,
-    axis_context: AxisContext,
-    name_stack,
-    replicated_args=None,
-    arg_shardings=None,
-    result_shardings=None,
-):
-    """Lowers a top-level jaxpr to an MHLO module.
-
-    Handles the quirks of the argument/return value passing conventions of the
-    runtime.
-
-    This function has been modified from its original form in the JAX project at
-    https://github.com/google/jax/blob/c4d590b1b640cc9fcfdbe91bf3fe34c47bcde917/jax/interpreters/mlir.py#L625version
-    released under the Apache License, Version 2.0, with the following copyright notice:
-
-    Copyright 2021 The JAX Authors.
-    """
-
-    if any(lowerable_effects.filter_not_in(jaxpr.effects)):  # pragma: no cover
-        raise ValueError(f"Cannot lower jaxpr with effects: {jaxpr.effects}")
-
-    assert platform == "cpu"
-    assert arg_shardings is None
-    assert result_shardings is None
-
-    # MHLO channels need to start at 1
-    channel_iter = 1
-    # Create a keepalives list that will be mutated during the lowering.
-    keepalives = []
-    host_callbacks = []
-    ctx = ModuleContext(
-        None, platform, axis_context, name_stack, keepalives, channel_iter, host_callbacks
-    )
-    ctx.context.allow_unregistered_dialects = True
-    with ctx.context, ir.Location.unknown(ctx.context):
-        # register_dialect()
-        # Remove module name characters that XLA would alter. This ensures that
-        # XLA computation preserves the module name.
-        module_name = _module_name_regex.sub("_", module_name)
-        ctx.module.operation.attributes["sym_name"] = ir.StringAttr.get(module_name)
-        lower_jaxpr_to_fun(
-            ctx,
-            func_name,
-            jaxpr,
-            effects,
-            public=True,
-            create_tokens=True,
-            replace_tokens_with_dummy=True,
-            replicated_args=replicated_args,
-            arg_shardings=arg_shardings,
-            result_shardings=result_shardings,
-        )
-
-        for op in ctx.module.body.operations:
-            func_name = str(op.name)
-            is_entry_point = func_name.startswith('"jit_')
-            if is_entry_point:
-                continue
-            op.attributes["llvm.linkage"] = ir.Attribute.parse("#llvm.linkage<internal>")
-
-    return ctx.module, ctx.context
 
 
 def new_inner_tracer(trace: DynamicJaxprTrace, aval) -> DynamicJaxprTracer:
@@ -623,13 +485,6 @@ def input_type_to_tracers(
     """Creates an expanded list of tracers representing an input values of a Jaxpr program"""
     in_aval, _ = unzip2(in_type)
     return _input_type_to_tracers(maker, in_aval)
-
-
-# def get_referent_frame(frame):
-#     """Find referents in the specific frame"""
-#     # TODO: clarify the logic! Can we need other frames?
-#     val = frame.constvar_to_val.get(frame.tracer_to_var.get(id(self)))
-#     return self if val is None else get_referent_frame(val)
 
 
 def output_type_to_tracers(
@@ -1036,133 +891,3 @@ class DynshapePrimitive(Primitive):
         eqn = new_jaxpr_eqn(invars, outvars, self, params, [], source_info)
         trace.frame.add_eqn(eqn)
         return out_tracers if self.multiple_results else out_tracers.pop()
-
-
-def _gather_shape_rule_dynamic(
-    operand,
-    indices,
-    *,
-    dimension_numbers,
-    slice_sizes,
-    unique_indices,
-    indices_are_sorted,
-    mode,
-    fill_value,
-):  # pragma: no cover
-    """Validates the well-formedness of the arguments to Gather. Compared to the original version,
-    this implementation skips static shape checks if variable dimensions are used.
-
-    This function has been modified from its original form in the JAX project at
-    https://github.com/google/jax/blob/88a60b808c1f91260cc9e75b9aa2508aae5bc9f9/jax/_src/lax/slicing.py#L1438
-    version released under the Apache License, Version 2.0, with the following copyright notice:
-
-    Copyright 2021 The JAX Authors.
-    """
-    # pylint: disable=unused-argument
-    # pylint: disable=too-many-branches
-    # pylint: disable=consider-using-enumerate
-    # pylint: disable=chained-comparison
-    offset_dims = dimension_numbers.offset_dims
-    collapsed_slice_dims = dimension_numbers.collapsed_slice_dims
-    start_index_map = dimension_numbers.start_index_map
-
-    # Note: in JAX, index_vector_dim is always computed as below, cf. the
-    # documentation of the GatherDimensionNumbers class.
-    index_vector_dim = _rank(indices) - 1
-
-    # This case should never happen in JAX, due to the implicit construction of
-    # index_vector_dim, but is included for completeness.
-    if _rank(indices) < index_vector_dim or index_vector_dim < 0:
-        raise TypeError(
-            f"Gather index leaf dimension must be within [0, rank("
-            f"indices) + 1). rank(indices) is {_rank(indices)} and "
-            f"gather index leaf dimension is {index_vector_dim}."
-        )
-
-    # Start ValidateGatherDimensions
-    # In the error messages output by XLA, "offset_dims" is called "Output window
-    # dimensions" in error messages. For consistency's sake, our error messages
-    # stick to "offset_dims".
-    _is_sorted(offset_dims, "gather", "offset_dims")
-    _no_duplicate_dims(offset_dims, "gather", "offset_dims")
-
-    output_offset_dim_count = len(offset_dims)
-    output_shape_rank = len(offset_dims) + _rank(indices) - 1
-
-    for i in range(output_offset_dim_count):
-        offset_dim = offset_dims[i]
-        if offset_dim < 0 or offset_dim >= output_shape_rank:
-            raise TypeError(
-                f"Offset dimension {i} in gather op is out of bounds; "
-                f"got {offset_dim}, but should have been in "
-                f"[0, {output_shape_rank})"
-            )
-
-    if len(start_index_map) != indices.shape[index_vector_dim]:
-        raise TypeError(
-            f"Gather op has {len(start_index_map)} elements in "
-            f"start_index_map and the bound of dimension "
-            f"{index_vector_dim=} of indices is "
-            f"{indices.shape[index_vector_dim]}. These two "
-            f"numbers must be equal."
-        )
-
-    for i in range(len(start_index_map)):
-        operand_dim_for_start_index_i = start_index_map[i]
-        if operand_dim_for_start_index_i < 0 or operand_dim_for_start_index_i >= _rank(operand):
-            raise TypeError(
-                f"Invalid start_index_map; domain is "
-                f"[0, {_rank(operand)}), got: "
-                f"{i}->{operand_dim_for_start_index_i}."
-            )
-
-    _no_duplicate_dims(start_index_map, "gather", "start_index_map")
-
-    # _is_sorted and _sorted_dims_in_range are checked in the opposite order
-    # compared to the XLA implementation. In cases when the input is not sorted
-    # AND there are problematic collapsed_slice_dims, the error message will thus
-    # be different.
-    _is_sorted(collapsed_slice_dims, "gather", "collapsed_slice_dims")
-    _sorted_dims_in_range(collapsed_slice_dims, _rank(operand), "gather", "collapsed_slice_dims")
-    _no_duplicate_dims(collapsed_slice_dims, "gather", "collapsed_slice_dims")
-    # End ValidateGatherDimensions
-
-    if _rank(operand) != len(slice_sizes):
-        raise TypeError(
-            f"Gather op must have one slice size for every input "
-            f"dimension; got: len(slice_sizes)={len(slice_sizes)}, "
-            f"input_shape.rank={_rank(operand)}"
-        )
-
-    if len(slice_sizes) != len(offset_dims) + len(collapsed_slice_dims):
-        raise TypeError(
-            f"All components of the offset index in a gather op must "
-            f"either be a offset dimension or explicitly collapsed; "
-            f"got len(slice_sizes)={len(slice_sizes)}, "
-            f"output_slice_sizes={offset_dims}, collapsed_slice_dims="
-            f"{collapsed_slice_dims}."
-        )
-
-    # This section contains a patch suggested to the upstream.
-    for i in range(len(slice_sizes)):
-        slice_size = slice_sizes[i]
-        corresponding_input_size = operand.shape[i]
-
-        if jax.core.is_constant_dim(corresponding_input_size):
-            if not (slice_size >= 0 and corresponding_input_size >= slice_size):
-                raise TypeError(
-                    f"Slice size at index {i} in gather op is out of range, "
-                    f"must be within [0, {corresponding_input_size} + 1), "
-                    f"got {slice_size}."
-                )
-
-    for i in range(len(collapsed_slice_dims)):
-        bound = slice_sizes[collapsed_slice_dims[i]]
-        if bound != 1:
-            raise TypeError(
-                f"Gather op can only collapse slice dims with bound 1, "
-                f"but bound is {bound} for index "
-                f"{collapsed_slice_dims[i]} at position {i}."
-            )
-
-    return _gather_shape_computation(indices, dimension_numbers, slice_sizes)
