+++ frontend/catalyst/jax_tracer.py
@@ -27,6 +27,33 @@ from pennylane.operation import AnyWires, Operation, Wires
 from pennylane.tape import QuantumTape
 
 import catalyst
+from catalyst.jax_extras import (
+    ClosedJaxpr,
+    DynamicJaxprTrace,
+    DynamicJaxprTracer,
+    ExpansionStrategy,
+    InputSignature,
+    OutputSignature,
+    PyTreeDef,
+    PyTreeRegistry,
+    ShapedArray,
+    _abstractify,
+    _input_type_to_tracers,
+    cond_expansion_strategy,
+    convert_element_type,
+    deduce_avals,
+    deduce_signatures,
+    eval_jaxpr,
+    input_type_to_tracers,
+    jaxpr_remove_implicit,
+    jaxpr_to_mlir,
+    make_jaxpr2,
+    sort_eqns,
+    tree_flatten,
+    tree_structure,
+    tree_unflatten,
+    wrap_init,
+)
 from catalyst.jax_primitives import (
     AbstractQreg,
     compbasis_p,
@@ -54,33 +81,6 @@ from catalyst.jax_primitives import (
 )
 from catalyst.utils.contexts import EvaluationContext, EvaluationMode, JaxTracingContext
 from catalyst.utils.exceptions import CompileError
-from catalyst.utils.jax_extras import (
-    ClosedJaxpr,
-    DynamicJaxprTrace,
-    DynamicJaxprTracer,
-    ExpansionStrategy,
-    InputSignature,
-    OutputSignature,
-    PyTreeDef,
-    PyTreeRegistry,
-    ShapedArray,
-    _abstractify,
-    _input_type_to_tracers,
-    cond_expansion_strategy,
-    convert_element_type,
-    deduce_avals,
-    deduce_signatures,
-    eval_jaxpr,
-    input_type_to_tracers,
-    jaxpr_remove_implicit,
-    jaxpr_to_mlir,
-    make_jaxpr2,
-    sort_eqns,
-    tree_flatten,
-    tree_structure,
-    tree_unflatten,
-    wrap_init,
-)
 
 
 class Function:
