+++ frontend/catalyst/utils/contexts.py
@@ -32,8 +32,8 @@ from jax._src.source_info_util import reset_name_stack
 from jax.core import find_top_trace
 from pennylane.queuing import QueuingManager
 
+from catalyst.jax_extras import new_dynamic_main2
 from catalyst.utils.exceptions import CompileError
-from catalyst.utils.jax_extras import new_dynamic_main2
 
 
 class EvaluationMode(Enum):
